Extract color themes
