+++ Makefile.am
@@ -5,6 +5,7 @@ bbs_SOURCES = \
 	browser.c \
 	bbsrc.c \
 	color.c \
+	color_themes.c \
 	config.c \
 	edit.c \
 	filter.c \
@@ -44,7 +45,7 @@ test_inkey_LDADD = $(CMOCKA_LIBS)
 test_filter_SOURCES = test/test_filter.c test/test_helpers.c browser.c filter.c queue.c slist.c global.c
 test_filter_CFLAGS = $(AM_CFLAGS) $(CMOCKA_CFLAGS)
 test_filter_LDADD = $(CMOCKA_LIBS)
-test_color_SOURCES = test/test_color.c test/test_helpers.c color.c slist.c global.c
+test_color_SOURCES = test/test_color.c test/test_helpers.c color.c color_themes.c slist.c global.c
 test_color_CFLAGS = $(AM_CFLAGS) $(CMOCKA_CFLAGS)
 test_color_LDADD = $(CMOCKA_LIBS)
 test_config_SOURCES = test/test_config.c test/test_helpers.c config.c slist.c global.c
